backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
prisma>=0.15.0
python-dotenv>=1.0.0
sse-starlette>=2.1.0
httpx>=0.27.0
pydantic>=2.8.0
